+++ src/main/java/no/uio/ifi/trackfind/backend/repositories/MappingRepository.java
@@ -8,7 +8,7 @@ import java.math.BigInteger;
 import java.util.Collection;
 
 @Repository
-public interface MappingRepository extends JpaRepository<Mapping, BigInteger> {
+public interface MappingRepository extends JpaRepository<Mapping, Long> {
 
     Collection<Mapping> findByRepository(String repository);
